+++ build_runner/lib/src/entrypoint/run_script.dart
@@ -138,7 +138,6 @@ class RunCommand extends BuildRunnerCommand {
           );
       var result = await build(
         builderApplications,
-        deleteFilesByDefault: options.deleteFilesByDefault,
         enableLowResourcesMode: options.enableLowResourcesMode,
         configKey: options.configKey,
         buildDirs: buildDirs,